Refatorar constantes e utils do App
