+++ App.js
@@ -35,7 +35,6 @@ import {
   endOfMonth,
   endOfWeek,
   format,
-  getWeeksInMonth,
   startOfMonth,
   startOfWeek,
 } from 'date-fns';
@@ -51,37 +50,22 @@ import {
   saveUserSettings,
 } from './storage';
 import AddHabitSheet from './components/AddHabitSheet';
-
-// --- IMAGENS ANIMADAS PARA OS MESES ---
-const MONTH_IMAGES = [
-  require('./assets/months/jan.gif'),
-  require('./assets/months/feb.gif'),
-  require('./assets/months/mar.gif'),
-  require('./assets/months/apr.gif'),
-  require('./assets/months/may.gif'),
-  require('./assets/months/jun.gif'),
-  require('./assets/months/jul.gif'),
-  require('./assets/months/aug.gif'),
-  require('./assets/months/sep.gif'),
-  require('./assets/months/oct.gif'),
-  require('./assets/months/nov.gif'),
-  require('./assets/months/dec.gif'),
-];
-
-const MONTH_NAMES = [
-  'JANEIRO', 'FEVEREIRO', 'MARÇO', 'ABRIL', 'MAIO', 'JUNHO',
-  'JULHO', 'AGOSTO', 'SETEMBRO', 'OUTUBRO', 'NOVEMBRO', 'DEZEMBRO'
-];
-
-const getMonthImageSource = (monthIndex, customImages) => {
-  const index = monthIndex % 12;
-
-  if (customImages && customImages[index]) {
-    return { uri: customImages[index] };
-  }
-
-  return MONTH_IMAGES[index];
-};
+import { MONTH_NAMES, getMonthImageSource } from './constants/months';
+import { DEFAULT_USER_SETTINGS } from './constants/userSettings';
+import { LEFT_TABS, RIGHT_TABS, getNavigationBarThemeForTab } from './constants/navigation';
+import { interpolateHexColor, lightenColor } from './utils/colorUtils';
+import { getDateKey, getMonthId, getMonthStart, shouldTaskAppearOnDate } from './utils/dateUtils';
+import { clamp01, clampValue } from './utils/mathUtils';
+import {
+  getQuantumProgressLabel,
+  getQuantumProgressPercent,
+  getSubtaskCompletionStatus,
+  getTaskCompletionStatus,
+  getTaskTagDisplayLabel,
+  normalizeTaskTagKey,
+} from './utils/taskUtils';
+import { formatTaskTime, toMinutes } from './utils/timeUtils';
+import { buildWavePath } from './utils/waveUtils';
 
 // --- COMPONENTE DA FAIXA DO TOPO ---
 const StickyMonthHeader = ({ date, customImages }) => {
@@ -296,440 +280,6 @@ function CustomizeCalendarModal({ visible, onClose, customImages, onUpdateImage 
 }
 
 
-const LEFT_TABS = [
-  {
-    key: 'today',
-    label: 'Today',
-    icon: 'time-outline',
-  },
-  {
-    key: 'calendar',
-    label: 'Calendar',
-    icon: 'calendar-clear-outline',
-  },
-];
-
-const RIGHT_TABS = [
-  {
-    key: 'discover',
-    label: 'Discover',
-    icon: 'compass-outline',
-  },
-  {
-    key: 'profile',
-    label: 'Profile',
-    icon: 'person-outline',
-  },
-];
-
-const NAV_BAR_THEMES = {
-  today: {
-    buttonStyle: 'dark',
-  },
-  calendar: {
-    buttonStyle: 'dark',
-  },
-  discover: {
-    buttonStyle: 'dark',
-  },
-  profile: {
-    buttonStyle: 'dark',
-  },
-};
-
-const DEFAULT_NAV_BAR_THEME = NAV_BAR_THEMES.calendar;
-
-const getNavigationBarThemeForTab = (tabKey) => NAV_BAR_THEMES[tabKey] ?? DEFAULT_NAV_BAR_THEME;
-
-const DEFAULT_USER_SETTINGS = {
-  activeTab: 'today',
-  selectedTagFilter: 'all',
-};
-
-const getDateKey = (date) => {
-  const normalized = new Date(date);
-  normalized.setHours(0, 0, 0, 0);
-  const year = normalized.getFullYear();
-  const month = String(normalized.getMonth() + 1).padStart(2, '0');
-  const day = String(normalized.getDate()).padStart(2, '0');
-  return `${year}-${month}-${day}`;
-};
-
-const getMonthStart = (date) => {
-  const normalized = new Date(date);
-  normalized.setHours(0, 0, 0, 0);
-  normalized.setDate(1);
-  return normalized;
-};
-
-const getMonthId = (date) => {
-  const normalized = getMonthStart(date);
-  return `${normalized.getFullYear()}-${String(normalized.getMonth() + 1).padStart(2, '0')}`;
-};
-
-const calculateWeeksInMonth = (date) => {
-  try {
-    if (typeof getWeeksInMonth === 'function') {
-      return getWeeksInMonth(date, { weekStartsOn: 0 });
-    }
-  } catch (error) {
-    // Fallback to manual calculation below
-  }
-
-  const start = startOfWeek(startOfMonth(date));
-  const end = endOfWeek(endOfMonth(date));
-  const days = (end.getTime() - start.getTime()) / (1000 * 60 * 60 * 24) + 1;
-  return Math.round(days / 7);
-};
-
-const hexToRgb = (hex) => {
-  const sanitized = hex.replace('#', '');
-  const bigint = parseInt(sanitized, 16);
-  return {
-    r: (bigint >> 16) & 255,
-    g: (bigint >> 8) & 255,
-    b: bigint & 255,
-  };
-};
-
-const lightenColor = (hex, amount = 0.7) => {
-  try {
-    const { r, g, b } = hexToRgb(hex);
-    const mixChannel = (channel) => Math.round(channel + (255 - channel) * amount);
-    const mixed = [mixChannel(r), mixChannel(g), mixChannel(b)];
-    return `#${mixed.map((channel) => channel.toString(16).padStart(2, '0')).join('')}`;
-  } catch (error) {
-    return '#f2f3f8';
-  }
-};
-
-const formatNumber = (value) => value.toString().padStart(2, '0');
-
-const formatTimeValue = ({ hour, minute, meridiem }) =>
-  `${formatNumber(hour)}:${formatNumber(minute)} ${meridiem}`;
-
-const toMinutes = ({ hour, minute, meridiem }) => {
-  const normalizedHour = meridiem === 'PM' ? (hour % 12) + 12 : hour % 12;
-  return normalizedHour * 60 + minute;
-};
-
-const formatTaskTime = (time) => {
-  if (!time || !time.specified) {
-    return 'Anytime';
-  }
-
-  if (time.mode === 'period' && time.period) {
-    const { start, end } = time.period;
-    return `${formatTimeValue(start)} - ${formatTimeValue(end)}`;
-  }
-
-  if (time.point) {
-    return formatTimeValue(time.point);
-  }
-
-  return 'Anytime';
-};
-
-const clampValue = (value, min, max) => Math.min(Math.max(value, min), max);
-
-const formatDuration = (totalSeconds) => {
-  const safeSeconds = Math.max(0, totalSeconds || 0);
-  const minutes = Math.floor(safeSeconds / 60);
-  const seconds = safeSeconds % 60;
-  return `${minutes}:${String(seconds).padStart(2, '0')}`;
-};
-
-const getQuantumProgressLabel = (task) => {
-  if (!task || task.type !== 'quantum' || !task.quantum) {
-    return null;
-  }
-  const mode = task.quantum.mode;
-  if (mode === 'timer') {
-    const minutes = task.quantum.timer?.minutes ?? 0;
-    const seconds = task.quantum.timer?.seconds ?? 0;
-    const limitSeconds = minutes * 60 + seconds;
-    if (!limitSeconds) {
-      return null;
-    }
-    const doneSeconds =
-      typeof task.quantum.doneSeconds === 'number'
-        ? task.quantum.doneSeconds
-        : task.completed
-        ? limitSeconds
-        : 0;
-    return `${formatDuration(doneSeconds)}/${formatDuration(limitSeconds)}`;
-  }
-  if (mode === 'count') {
-    const limitValue = task.quantum.count?.value ?? 0;
-    if (!limitValue) {
-      return null;
-    }
-    const unit = task.quantum.count?.unit?.trim() ?? '';
-    const doneValue =
-      typeof task.quantum.doneCount === 'number'
-        ? task.quantum.doneCount
-        : task.completed
-        ? limitValue
-        : 0;
-    return `${doneValue}/${limitValue}${unit ? ` ${unit}` : ''}`;
-  }
-  return null;
-};
-
-const clamp01 = (value) => Math.min(1, Math.max(0, value));
-
-const getQuantumProgressPercent = (task) => {
-  if (!task || task.type !== 'quantum' || !task.quantum) {
-    return 0;
-  }
-  const mode = task.quantum.mode;
-  if (mode === 'timer') {
-    const minutes = task.quantum.timer?.minutes ?? 0;
-    const seconds = task.quantum.timer?.seconds ?? 0;
-    const totalSeconds = minutes * 60 + seconds;
-    if (!totalSeconds) {
-      return 0;
-    }
-    const doneSeconds =
-      typeof task.quantum.doneSeconds === 'number' ? task.quantum.doneSeconds : 0;
-    return clamp01(doneSeconds / totalSeconds);
-  }
-
-  if (mode === 'count') {
-    const limitValue = task.quantum.count?.value ?? 0;
-    if (!limitValue) {
-      return 0;
-    }
-    const doneCount = typeof task.quantum.doneCount === 'number' ? task.quantum.doneCount : 0;
-    return clamp01(doneCount / limitValue);
-  }
-
-  return 0;
-};
-
-const mixChannel = (from, to, ratio) => Math.round(from + (to - from) * ratio);
-
-const interpolateHexColor = (from, to, ratio) => {
-  const normalize = (hex) => hex.replace('#', '');
-  const fromValue = parseInt(normalize(from), 16);
-  const toValue = parseInt(normalize(to), 16);
-  const fromRgb = {
-    r: (fromValue >> 16) & 255,
-    g: (fromValue >> 8) & 255,
-    b: fromValue & 255,
-  };
-  const toRgb = {
-    r: (toValue >> 16) & 255,
-    g: (toValue >> 8) & 255,
-    b: toValue & 255,
-  };
-  const mixed = {
-    r: mixChannel(fromRgb.r, toRgb.r, ratio),
-    g: mixChannel(fromRgb.g, toRgb.g, ratio),
-    b: mixChannel(fromRgb.b, toRgb.b, ratio),
-  };
-  return `#${[mixed.r, mixed.g, mixed.b].map((value) => value.toString(16).padStart(2, '0')).join('')}`;
-};
-
-const buildWavePath = ({ width, height, amplitude, phase }) => {
-  if (!width || !height) {
-    return '';
-  }
-  const points = 24;
-  const step = width / points;
-  const center = height * 0.5;
-  let path = `M 0 ${center}`;
-  for (let i = 0; i <= points; i += 1) {
-    const x = step * i;
-    const theta = (i / points) * Math.PI * 2 + phase;
-    const y = center + Math.sin(theta) * amplitude;
-    path += ` L ${x.toFixed(2)} ${y.toFixed(2)}`;
-  }
-  path += ` L ${width} ${height}`;
-  path += ` L 0 ${height} Z`;
-  return path;
-};
-
-const getTaskCompletionStatus = (task, date) => {
-  if (!task || !date) {
-    return false;
-  }
-
-  const dateKey = typeof date === 'string' ? date : getDateKey(date);
-  if (!dateKey) {
-    return false;
-  }
-
-  if (task.completedDates && typeof task.completedDates === 'object') {
-    return Boolean(task.completedDates[dateKey]);
-  }
-
-  if (!task.repeat && task.dateKey) {
-    return task.dateKey === dateKey ? Boolean(task.completed) : false;
-  }
-
-  return false;
-};
-
-const getSubtaskCompletionStatus = (subtask, dateKey) => {
-  if (!subtask) {
-    return false;
-  }
-
-  if (dateKey && subtask.completedDates && typeof subtask.completedDates === 'object') {
-    if (subtask.completedDates[dateKey] === true) {
-      return true;
-    }
-    if (subtask.completedDates[dateKey] === false) {
-      return false;
-    }
-  }
-
-  if (dateKey) {
-    return false;
-  }
-
-  return Boolean(subtask.completed);
-};
-
-const normalizeTaskTagKey = (task) => {
-  if (!task) {
-    return null;
-  }
-  if (task.tag && typeof task.tag === 'string') {
-    const normalized = task.tag.trim();
-    if (!normalized || normalized.toLowerCase() === 'none' || normalized.toLowerCase() === 'no_tag') {
-      return null;
-    }
-    return normalized;
-  }
-  if (task.tagLabel && typeof task.tagLabel === 'string') {
-    const label = task.tagLabel.trim();
-    if (!label || label.toLowerCase() === 'no tag') {
-      return null;
-    }
-    return label
-      .toLowerCase()
-      .replace(/[^a-z0-9]+/g, '_')
-      .replace(/^_|_$/g, '');
-  }
-  return null;
-};
-
-const getTaskTagDisplayLabel = (task) => {
-  if (!task) {
-    return null;
-  }
-  if (task.tagLabel && typeof task.tagLabel === 'string') {
-    const label = task.tagLabel.trim();
-    if (!label || label.toLowerCase() === 'no tag') {
-      return null;
-    }
-    return label;
-  }
-  if (task.tag && typeof task.tag === 'string') {
-    const normalized = task.tag.trim();
-    if (!normalized || normalized.toLowerCase() === 'none' || normalized.toLowerCase() === 'no_tag') {
-      return null;
-    }
-    return normalized
-      .split('_')
-      .filter(Boolean)
-      .map((segment) => segment.charAt(0).toUpperCase() + segment.slice(1))
-      .join(' ');
-  }
-  return null;
-};
-
-const normalizeDateValue = (value) => {
-  if (!value) {
-    return null;
-  }
-  if (value instanceof Date) {
-    const date = new Date(value);
-    date.setHours(0, 0, 0, 0);
-    return date;
-  }
-  if (typeof value === 'string' && /^\d{4}-\d{2}-\d{2}$/.test(value)) {
-    const [year, month, day] = value.split('-').map((part) => Number.parseInt(part, 10));
-    const date = new Date(year, month - 1, day);
-    date.setHours(0, 0, 0, 0);
-    return date;
-  }
-  const date = new Date(value);
-  if (Number.isNaN(date.getTime())) {
-    return null;
-  }
-  date.setHours(0, 0, 0, 0);
-  return date;
-};
-
-const isSameDay = (dateA, dateB) => {
-  if (!dateA || !dateB) {
-    return false;
-  }
-  return dateA.getTime() === dateB.getTime();
-};
-
-const getWeekdayKeyFromDate = (date) => {
-  const WEEKDAY_KEYS = ['sun', 'mon', 'tue', 'wed', 'thu', 'fri', 'sat'];
-  return WEEKDAY_KEYS[date.getDay()] ?? null;
-};
-
-const shouldTaskAppearOnDate = (task, targetDate) => {
-  if (!task || !targetDate) {
-    return false;
-  }
-
-  const normalizedTargetDate = normalizeDateValue(targetDate);
-  const normalizedStartDate = normalizeDateValue(task.date ?? task.dateKey);
-
-  if (!normalizedTargetDate || !normalizedStartDate) {
-    return false;
-  }
-
-  if (isSameDay(normalizedStartDate, normalizedTargetDate)) {
-    return true;
-  }
-
-  const repeat = task.repeat;
-  if (!repeat || !repeat.option || repeat.option === 'off') {
-    return false;
-  }
-
-  if (normalizedTargetDate.getTime() < normalizedStartDate.getTime()) {
-    return false;
-  }
-
-  switch (repeat.option) {
-    case 'daily':
-      return true;
-    case 'weekly':
-      return normalizedTargetDate.getDay() === normalizedStartDate.getDay();
-    case 'monthly':
-      return normalizedTargetDate.getDate() === normalizedStartDate.getDate();
-    case 'weekend': {
-      const day = normalizedTargetDate.getDay();
-      return day === 0 || day === 6;
-    }
-    case 'weekdays': {
-      const day = normalizedTargetDate.getDay();
-      return day >= 1 && day <= 5;
-    }
-    case 'custom': {
-      const weekdays = Array.isArray(repeat.weekdays) ? repeat.weekdays : [];
-      if (!weekdays.length) {
-        return false;
-      }
-      const weekdayKey = getWeekdayKeyFromDate(normalizedTargetDate);
-      return weekdayKey ? weekdays.includes(weekdayKey) : false;
-    }
-    default:
-      return false;
-  }
-};
-
-// --- COMPONENTE ATUALIZADO: RELATÓRIO DO DIA COM GIF ---
 function DayReportModal({ visible, date, tasks, onClose, customImages }) {
   const { height } = useWindowDimensions();
