+++ constants/userSettings.js
@@ -1,0 +1,6 @@
+const DEFAULT_USER_SETTINGS = {
+  activeTab: 'today',
+  selectedTagFilter: 'all',
+};
+
+export { DEFAULT_USER_SETTINGS };
